conf: Move virDomainPCIAddressBusIsFullyReserved()
This function was private to the QEMU driver and was,
accordingly, called qemuDomainPCIBusFullyReserved().

However, since it's really not QEMU-specific at all,
it makes a lot of sense to have it closer to the
virDomainPCIAddressBus struct it operates on; even more
so considering we're going to introduce a complementary
virDomainPCIAddressBusIsEmpty() function later on.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <1498747617-5624-1-git-send-email-abologna@redhat.com>
Reviewed-by: Laine Stump <laine@laine.org>