virFileInData: Report an error if unable to reposition file
The purpose of this function is to tell if the current position
in given FD is in data section or a hole and how much bytes there
is remaining until the end of the section. This is achieved by
couple of lseeks(). The most important part is that we reposition
the FD back, so that the position is unchanged from the caller
POV. And until now the final lseek() back to the original
position was done with no check for errors. And I was convinced
that that's okay since nothing can go wrong. However, John
persuaded me, that it's better to be safe than sorry. Therefore,
lets check if the final lseek() succeeded and if it doesn't
report an error.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <ef81e1dc34c10af0e90a4545fb342691889ab638.1499763929.git.mprivozn@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>