libxl: fix cdrom default driver name
In recent libvirt libxl does fill non-existant driver elements with type='raw',
but lacks the name='qemu' attribute that it had in the past.
Without that the xml fails to validate, therefore old XML files can't be
edited/defined anymore.

Since the cdrom device type only works with the qemu driver, add that as
the default attribute just as it was added in the past.

Example of the verification error:
  $ virt-xml-validate mytest.xml
  Relax-NG validity error : Extra element devices in interleave

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Message-Id: <1500302574-4012-1-git-send-email-christian.ehrhardt@canonical.com>