nodedev: Fix usage of virNodeDeviceObjListGetParentHost
In order to avoid an eventual possible race, modify the call to not
use the @def, but rather take the various fields needed. The race could
occur because the Destroy path needs to Unlock (and Unref) the object.
This could lead to an eventual change event from udevAddOneDevice which
would free obj->def and it's fields that this code would be attempting
to reference. So better safe than sorry.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170717150224.2988-2-jferlan@redhat.com>