nwfilter: Add @refs logic to __virNWFilterObj
"Simple" conversion to the virObjectLockable object isn't quite possible
yet because the nwfilter lock requires usage of recursive locking due to
algorithms handing filter "<rule>"'s and "<filterref>"'s. The list search
logic would also benefit from using hash tables for lookups. So this patch
is step 1 in the process - add the @refs to _virNWFilterObj and modify the
algorithms to use (a temporary) virNWFilterObj{Ref|Unref} in order to set
things up for the list logic to utilize virObjectLockable hash tables.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170718205750.14503-2-jferlan@redhat.com>