nwfilter: Remove recursive locking for nwfilter instantiation
The current algorithm required usage of recursive locks because
there was no other mechanism available to ensure that the object
wasn't deleted whilst the instantiation was taking place.

Unfortunately common objects (virObjectLockable) do not allow for
recursive locks; however, since this is a very specific use case,
let's roll our own version by using pthread_mutex_trylock and handle
the return status based on what the API tells us.

In order to do this, introduce a virMutexTryLock call and modify the
virNWFilterObjListFindInstantiateFilter in order to consume and check
the returned status. For starters let's only ever worry about the
good status (e.g. we were the first to get the lock) and the pseudo
recursive view of the world that our current thread was the thread
that tried to get the lock when EBUSY is returned. An EBUSY is
returned when either this thread or some thread has the lock. This
involves keeping track of which thread was able to take out the first
lock and if it's the same as the current thread, then allow the lock
to succeed; otherwise, we'll need to retry as some other thread
is currently holding the lock. Still we don't want that retry to last
forever, so in order to avoid potential deadlock if two threads are
doing a define at the same time, add retry lock that lasts for about
2 seconds which should be plenty of time.For any other errors, cause
failure for the calling thread.

Modify the callers to use a new virNWFilterObjEndInstAPI which handles
the comparison of the number of try locks that were successful for
the thread and only truly unlocks the object resource when the last
lock is removed.

Additionally to avoid the possibility that we do something untoward
to ourselves and because we are not taking out the higher level
lock, introduce an instLock to ensure that changes to the saved
depth and tid are single threaded.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170718205750.14503-5-jferlan@redhat.com>