nwfilter: Convert virNWFilterObj to use virObjectLockable
Now that we have a bit more control, let's convert our object into a
lockable object and let that magic handle the create, lock/unlock,
and reference counting.

Because we have the need for instantiation in a recursive manner,
introduce the virObjectTryLock to handle the virNWFilterObjTryLock
processing. It'll just be the shim into virMutexTryLock, but adds
an extra error value EFAULT for when the incoming @obj is not
determined to be a LockableObject.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170718205750.14503-6-jferlan@redhat.com>