numa: compute and set matching vcpus for numa domains
The QEMU driver can erroneously allocate more vpus to a domain
than there are cpus in the domain if the <numa> element is used
to describe <cpu> element topology. Fix this by calculating
the number of cpus described in the <numa> element of a <cpu>
element and comparing it to the number of vcpus. If the number
of vcpus is greater than the number of cpus, cap the number of
vcpus to the number of cpus.

This patch also adds a small libvirt documentation update under
the "CPU Allocation" section.

Signed-off-by: Wim ten Have <wim.ten.have@oracle.com>
Message-Id: <20170720092926.15951-2-wim.ten.have@oracle.com>