qemu: undefine is not allowed during domain starting up
Start a domain whilst undefine it, if starting failed duing ProcessLaunch,
on which period qemu exited unexpectedly, the operation will lead to failure
of undefine the domain until libvirtd restarted. The reason is that libvirtd
will unlock vm during qemuProcessStart, qemuDomainUndefineFlags can get the
lock and set vm->persistent 0 but not remove the "active" domain.

Signed-off-by: Yi Wang <wang.yi59@zte.com.cn>
Message-Id: <1500713749-46892-1-git-send-email-wang.yi59@zte.com.cn>