conf: Introduce the virDomainNetModel enumeration
Up until now, we have stored the model name for network
interfaces as raw strings in virDomainNetDef: this is
suboptimal for a number of reasons, such as having to
perform relatively expensive string allocations and
comparisons all the time and not giving the compiler
the opportunity to have our back in certain situations.

Replace the strings with an enumeration similar to the
ones we already use for pretty much everything else.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <1501071605-25282-3-git-send-email-abologna@redhat.com>