conf: check for buffer errors before virBufferUse
After an OOM error, virBuffer* APIs set buf->use to zero.
Adding a buffer to the parent buffer only if use is non-zero
would quitely drop data on error.

Check the error beforehand to make sure buf->use is zero
because we have not attempted to add anything to it.
Message-Id: <92409353e297a7199b4683a4b672d1d1400de829.1501075760.git.jtomko@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>