security: apparmor: load the storage driver dynamically
In commit 5e515b542d I've attempted to fix the inability to access
storage from the apparmor helper program by linking with the storage
driver. By linking with the .so the linker complains that it's not
portable. Fix this by loading the module dynamically as we are supposed
to do.
Message-Id: <8b0b294cdc89d331265d925cf776fb29a60887fd.1501099726.git.pkrempa@redhat.com>
Reviewed-by: Daniel P. Berrange <berrange@redhat.com>