conf: Fix printing of 'type' and 'tty_compat' for Chr devices
Commit id '0c1d8632' caused a regression in the virt-manager
test suite when formatting the <smartcard mode='passthrough'
type='spicevmc'/>.

Adust the code to print the type in it's own new helper called
virDomainChrTypeFormat and have the virDomainChrSourceDefFormat
manage just formatting the source and change to a void type since
only 0 could be returned. Adjust the callers to handle properly.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170802195134.28814-1-jferlan@redhat.com>