qemu: Honour <on_reboot/>
https://bugzilla.redhat.com/show_bug.cgi?id=1476866

For some reason, we completely ignore <on_reboot/> setting for
domains. The implementation is simply not there. It never was.
However, things are slightly more complicated. QEMU sends us two
RESET events on domain reboot. Fortunately, the event contains
this 'guest' field telling us who initiated the reboot. And since
we don't want to destroy the domain if the reset is initiated by
a user, we have to ignore those events. Whatever, just look at
the code.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <555a0dc2ba8b4c8cda3bdb66b7615fe6b9c64866.1501745787.git.mprivozn@redhat.com>